requirements.txt
streamlit
streamlit-pandas-profiling
pandas
st-gsheets-connection
ydata_profiling
numba==0.58.1
matplotlib.pyplot
plotly.express
seaborn